Allow channel parameter for remote control

+++ BlinkStickClient.Base/DataModel/CustomNotification.cs
@@ -45,7 +45,7 @@ namespace BlinkStickClient.DataModel
 
         public event EventHandler<PatternSendEventArgs> PatternSend;
 
-        protected void OnPatternSend(byte channel, byte ledFirst, byte ledLast, BlinkStickDeviceSettings device, Pattern pattern)
+        protected void OnPatternSend(int channel, int ledFirst, int ledLast, BlinkStickDeviceSettings device, Pattern pattern)
         {
             if (PatternSend != null)
             {
@@ -157,13 +157,13 @@ namespace BlinkStickClient.DataModel
 
     public class PatternSendEventArgs: EventArgs
     {
-        public byte Channel;
-        public byte LedFirst;
-        public byte LedLast;
+        public int Channel;
+        public int LedFirst;
+        public int LedLast;
         public BlinkStickDeviceSettings Device;
         public Pattern Pattern;
 
-        public PatternSendEventArgs(byte channel, byte ledFirst, byte ledLast, BlinkStickDeviceSettings device, Pattern pattern)
+        public PatternSendEventArgs(int channel, int ledFirst, int ledLast, BlinkStickDeviceSettings device, Pattern pattern)
         {
             this.Channel = channel;
             this.LedFirst = ledFirst;
